Remove ActorSystemRef from MailBox
Since the lifetime of the Registration, the otherside of the Notifier, is tied to the lifetime of the actor, which in turn is tied to the lifetime of the actor system it is in, it has no point to keep this reference around.

Basically if we can't notify the actor one of two things can happend:
1. actor is removed from the actor system, or
2. the actor system is shutdown.

It can't happen that the actor is still alive and the actor system is stopped. Which means the check if the actor system is alive after notifing the actor is pointless.