Rewrite Initiator trait
This splits the Initiator trait into two: NewInitiator and Initiator,
following NewActor and Actor.

This for example allows a TcpListener to bound on each thread, rather
then duplicate the original file descriptor. Although this may not seem
like much it will avoid the thundering herd problem, where a single
incoming connection will wake all worker threads.

Since TcpListener was the only implementation of the old Initiator trait
it's change is included in this commit.

A new type NewTcpListener is added that implements NewInitiator, it acts
as a intermediate representation of TcpListener.

Left to do is adding a setup function to NewTcpListener to set options
on TcpListener. Also testing if binding to "localhost:0" still works, or
if it will bind to a different port on each worker.