Add associative type Error to the NewActor trait
This also changes the new method to return a Result<Actor, Error>.
Because NewActor.new can now return a result a number of public
functions have also changed. ActorSystem.spawn and test::init_actor were
both changed to return a result.

The implementation of NewActor for asynchronous functions use the never
type (`!`) as error and thus never return an error. This means that a
lot of examples need to unwrap the results from calls to spawn new
actors.

This also changes the implementation of TcpListener and ActorProcess.
In ActorProcess when restarting an actor returns an error it now just
logs it and marks the process as complete, as not much else can be done.
This will needs improvement in the future, see below. The same goes for
TcpListener, it needs better error handling in its Initiator.poll
implementation.

Some things to improve in the future:
 * Better handling of NewActor errors in both ActorProcess and
   TcpListener, both now log and give up.
 * ActorProcess.run needs to be refactored, it's getting too messy.