Change the setup function to return a generic error
Rather then an io::Error.

The also adds a generic parameter SetupError to RuntimeError, defaulting
to the never type.

Since the examples make use of the RuntimeError type most of them were
changed as well. And while I was there I made some improvements,
including expanding the comments in the examples themselves.