Merge actor reference into one ActoRef struct
Using a single type rather then three makes it easier to use in
situations where the actual type of the actor reference doesn't really
matter.

The remote actor reference is removed as that wasn't implemented anyway
and should be much easier to add in the future. For the remaining two
types a type alias is added to keep the code working, as the API of the
types hasn't changed.

The design of the trait is a bit odd, but it has a reason. When typing I
don't want to repeat the `M` parameter, i.e. I want to type `ActorRef<M,
Local>` not `ActorRef<M, Local<M>>`. To achieve this an empty type, e.g.
`Local`, was added that implements `ActorRefType`. This type wouldn't
have any generic type parameters. The actual data, which needs to be
generic for a specific generic type parameter, is defined in another
struct which is used in the `ActorRefType` as associated type `Data`.
This is the data that is actually stored in an `ActorRef` and on this
data the `ActorRefType` operates.