Add SyncSupervisor trait
The Supervisor trait (for asynchronous actors) will in the future need
to handle errors returned by there NewActor implementation (after an
actor is restarted). Synchronous actors however don't have a way to be
"created" and only have a run function.

So this future version of Supervisor can't be used for synchronous
actors and it needs its own version, that is this trait.