Port tcp::Server to Mio
This changes the API to bind the TcpListener in tcp::Server::setup,
rather then when the NewActor implementation of tcp::ServerSetup is
called.

Gaea sets `SO_REUSEPORT` be default, but Mio doesn't and doesn't have a
convenient way to set it. Furthermore the option is somewhat hard to use
in a cross-platform way. But because we don't have the option set
anymore we can't bind(2) twice using the same port, so the old code
doesn't work anymore (when using more then one worker thread).