Handle worker thread errors better
Rather then join the worker threads in order this will wait until an
event is generated for the worker thread (by dropping the receiving side
of a Unix pipe).

The Unix in the future can be used to send signals to workers threads,
e.g. any processes signals received that the coordinator thread will
handle in the future.