Rewrite scheduler
There are two components to the new scheduler:

* `RunQueue`: holds the processes that are ready to run.
* `Tree`: holds the inactive processes.

The scheduler has unique access to the `Tree`. However the `RunQueue` is
shared with `WorkStealer`, which can be used to steal processes in the
ready state from this scheduler. This is used by other workers threads
to steal work for themselves in an effort to prevent an in-balance in
the workload.

The `RunQueue` is behind a `RwLock` to enable a fast path for the owner
of the run queue if no other thread is attempting to access it at the
same time. To ensure the thread that owns the run queue is never blocked
only the `Scheduler` can request mutable access, the `WorkStealer` may
not. The `WorkStealer` can only request immutable access, this way other
`WorkStealer`s and the `Scheduler` can continue to operate on the run
queue concurrently (albeit using the slow path), ensure the thread that
owns the run queue isn't blocked.

Processes can be in one of the following states:

* Inactive: default state of an process, its located in `Scheduler.inactive`.
* Ready: process is ready to run (after they are marked as such, see
  `Scheduler::mark_ready`), located in `Scheduler.ready`.
* Running: process is being run, located on the stack in
  `Scheduler::run_process` or `WorkStealer::steal_and_run`.
* Stopped: final state of a process, at this point its deallocated and its
  resources cleaned up.