Accept all queued connection in tcp::Server after receiving shutdown message
We're using `SO_REUSEPORT` which on most OSes causes each listener (file
descriptor) to have there own accept queue. This means that connections
in the queue would be dropped if we would close the file descriptor
immediately. So we first accept all pending connections and start actors
for them.

Note however that there is still a race condition between the last call
to `accept` and the time the file descriptor is actually closed,
currently we can't avoid this.