Fix Receiver::try_recv when concurrently dropping the last Sender
Before this change the Receiver would check the channel for message once
and afterwords check if it was still connected. If it wasn't connected
it would return RecvError::Disconnected.

However the was a race condition between the Receiver checking the
channel for messages and checking if it was still connected, with the
last Sender sending a message and dropping itself concurrently.

To fix this it now checks a second time for message before returning
RecvError::Disconnected, ensuring that the case above no longer causes
problems.