Reduce some of flakyness of threaded tests
Before the tests would fail running too many iterations. But for some
tests the amount of work done per iteration is so small that it
completes all iterations before its partner thread(s) have a chance to
run.

To improve this we put a lock in place, used by the start_threads macro,
to reduce the number of thread started/running concurrently. We also
increase the number of iterations in the loop macro.