Ensure that we adding back shared processes it correctly scheduled
There was a small race condition between the time the process was added
to the inactive list and the coordinator marking it as active. Now when
a process is added back its added to the inactive queue first, then it's
checked if an event was missed and moved to the run queue if so.