Fix shared newly added processes not being scheduled
Before this commit the process would blindly be added to the list of
inactive processes. However it could be possible that in between the
time e.g. a TcpStream for the process was registered and the process is
added to the scheduler the coordinator thread already received an event
for it. This commit changes the AddActor::add function to call
SchedulerRef::add_process which checks if the process should already be
marked as ready.

This causes roughly a delay of 1 month in my Msc thesis. If someone ever
reads this... Don't build a distributed system for your thesis, it's too
much work.