Add new Bytes trait
Trait to make easier to work with slices of (uninitialised) bytes.

This will be used in the network types in functions such as
TcpStream::try_recv, allowing them to be used by a bytes slice (like
io::Read::read), but also easily with Vectors and the user's own buffer
types.