Remove to_mark_ready set from shared Scheduler
Instead mark processes as ready-to-run in the Inactive list. This make
the Scheduler::mark_ready and SchedulerRef::add_process quite a bit
simpler.

When marking ready: instead of having to check if a process is in the
inactive queue, if not add a mark to_mark_ready and then check again, we
simply leave a mark in the Inactive list. This will be picked up by
SchedulerRef::add_process, by using Inactive::add which now returns the
process if its marked as ready-to-run. This remove the need for the
double checks and an additional HashSet.