Replace crossbeam_channel with inbox for ActorRefs
Reduces the memory usage from ~7.08 GB to 2.09 GB for the stress_memory
example. Furthermore it reduces its startup time from ~10 seconds to ~5.

The main difference is that the inbox crate uses a bounded channel,
currently of size 8, where as we were using Crossbeam's unbounded
channel. Note that using Crossbeam's bounded channel increased the
memory usage to 8.90 GB for the stress_memory example.