Drop the channel's messages after Receiver is dropped without a Manager
We do this to support the use case were the channel holds a
oneshot::Sender and the receiver of the oneshot channel is holding a
Sender to this channel.

Effectively this creates a cyclic drop dependency: Sender -> Channel ->
oneshot::Sender which blocks oneshot::Receiver::recv.

If the actor holding a Sender calls oneshot::Receiver::recv it will wait
for a response or until the oneshot::Sender is dropped, while the actor
is holding a Sender to this channel. However if this Receiver is dropped
it won't drop the oneshot::Sender without the channel. This causes
oneshot::Receiver::recv to wait forever, while holding a Sender.