Fix missing event from short running sync actors
If a sync actor is short running, or the time before calling
Runtime::start takes a while, the thread running the sync actor could be
done before the coordinator had a chance to register it. This causes the
coordinator to never receive an event for it on FreeBSD (it does receive
an event on macOS and Linux) and never remove it from the list of sync
actors, causing the runtime to run for ever.

The fix is fairly simple: after registering check if the sync worker is
still alive by trying to write zero bytes to its pipe.