Ignore EINVAL in TcpStream::connect
macOS sometimes returns EINVAL when the socket is not yet connected when
calling getpeername in the Connect Future. Since we're ensure that the
arguments passed to getpeername are valid it's safe to ignore this error
and threat as socket is not (yet) connected.

Also deflakes the connect_connection_refused test.