Add basic trace infrastructure
The current format is JSON (like) format with the following properties:
 * stream_id: allows events to be correlated, we'll use one id per
   thread.
 * description: human-readable description of the event.
 * start time.
 * elapsed time.
 * attributes of the event, if any.

Note that events can overlap within a stream, this means that one event
was a child event of a parent event. E.g. when running a process A,
which generates an event, it could generate another event will the
process A is running.