Change the trace timestamp to be based on an internal epoch
This gives us ~584 years of continues running before hitting the limits
of the u64 representation. But more importantly it reduces amount of
system calls in the timing of events from three to two as the wall-clock
time is no longer needed.

The still give the trace a wall-clock time it now starts with a metadata
object contains a single timestamp with a SystemTime in debug formatting
(which still needs improvement).