Use reference to Channel in RecvValue
Rather than a reference to the Receiver, which is itself a pointer to
the Channel. Same as commit 5232923403ffa014b840d52927659ecd23c08b66 but
for the receiving end.

This also implement Send and Sync for Channel. Because RecvValue now has
a reference to it, it's required to make RecvValue Send and Sync.