Improve waking mechanism of thread-safe actors
Previously it would send a message along a channel and wake the
coordinator. The coordinator thread would than have to read the message
and schedule the correct actor.

However this was inefficient as it would require the coordinator to
always wake up.

This commit changes the implementation to be specific to the shared
process, leaving the old implementation untouched (which is still being
used for thread-local actors).

The new implementation has direct access to the shared runtime internals
and schedules the process in the scheduler itself. This avoids all
communication with the coordinator thread, reducing its wake-ups.