Split Runtime into Runtime and rt::Setup
rt::Setup is the build pattern type to configure the Runtime. The
Runtime type can be used to spawn actors, receive process signals and
run arbitrary function on the worker threads.

Removes Runtime::with_setup, replaced with Runtime::run_on_workers that
allows an arbitrary number of function to be run.

This removes the generic parameter from Runtime, rt::Setup and
rt::Error.

Also removes the hack module in rt.