Move scheduling based on shared timers to workers
Instead of letting the coordinator schedule thread-safe actors based on
the shared timers this commit changes the workers to schedule the
thread-safe actors.

This does introduce the "Thundering herd problem" for the shared timers
if multiple worker threads have the same timeout (based on the shared
timers) and all attempt to schedule the same thread-safe actor because
of it.