Change Channel's slot to be a slice
Allocating the Channel becomes a bit more complex as we need to manually
the layout for it and can't use the convenience of the Box type.

This also has the side effect of making the pointers in Sender, Receiver
and Manager a "fat" pointer. This means it will contain the regular
pointer and a length attribute, taking up (2 x 8) 16 bytes on 64bit.

This increases the size of Sender, Receiver, Manager from 8 to 16 bytes,
SendValue from 32 to 40 bytes and Join from 24 to 32 bytes.

If this size increase becomes a problem, or if we ever want to decrease
the memory usage we can change the fat pointer into a regular pointer
again and using some bits to indicate the length of the channel. For
examples if we have four different capacities, it can be encoded into
just 2 bits. These two bits can ORed with the regular pointer and be
received later.

The upside of such an approach would be that the size is reduced from 16
to 8 bytes. The downside is that on each access to the channel the
reference has to be prepared by unpacking the length from the pointer.
And ensuring the length is correct at all times is an additional
maintenance burden.