Add new function
Allows the creation of an user specified capacity bounded channel. The
current implementation is limited to a capacity in the range 1..=29.

This limitation is because of the size of the `status` in the Channel.
As each slot needs two bits for its status and we need enough bits for
the receiver position the current limit is 29. The following must hold
true, where $n is the capacity of the channel:
2 ^ (64 - ($N * 2)) >= $N.