Change test to use all possible capacities
This touches nearly all tests, using the new with_all_capacities! macro
to run a test with all capacities possible.

With the 'stress_testing' feature enabled this will actually run all
capacities, with it disabled it will only run three capacities (minimum,
maximum and the small capacity) for performance.

For the stress_sending_fill test it only uses capacities that are a
power of two. The problem is the wrap around of the receiver position
with capacities that are **not** a power of two.

For example taking a capacity of 5. The maximum receiver position is 63
(6 bits), which is index 3. The next value will be 0 (with wrap around),
which is index 0. This means we've skipped index 4. The crate doesn't
guaranteed FIFO ordering so it's fine, but the test assumes it.