Add shared::Timers
This is roughly the same implementation as local::Timers, but uses locks
internally allowing the lock around the entire structure to be removed.
Since the locks, and the critical sections that hold them, are smaller
this should improve the performance.

This fixes the "Thundering herd problem" problem by considering the
shared timers only in a single worker at a time. That worker will than
process the shared timers and wakes other workers if needed.

This will create a wave of activity, multiple workers waking up to
access the shared scheduler, but at least they'll have something to do.