Add trace::CoordinatorLog
The is a special version of trace::Log for the coordinator. It also
adds trace::SharedLog which is the data shared between one
trace::CoordinatorLog and multiple trace::Logs, which includes the file.

It also includes the stream counter for the coordinator (0) stream. This
will be used in future commits to log events for thread-safe actors.