Add substream to trace format
This allows us to expose the tracing API to the async actors.
Specifically it allows us to create a substream for each actor, using
their process id.

In the convertion to Chrome's Trace Event Format we'll now start using
pids (process ids) for the threads and tid (thread ids) for the actors.
This way we don't have trace events that block each other, e.g. from two
actors who are multiplexed on the same worker thread.