Fix overlapping events in convert trace tool
Sometimes Instant returns a value that is equal to a previously returned
value. Catapult can't really deal with this and creates two overlapping
events, making them both unreadable and unusable.

To fix this we change the starting time and duration by a few
microseconds to ensure the two events don't start at the same time.