Fix updating of length in shared Inactive tree
In Branch::add_branches it would remove a process (replacing the pointer
with a branch) and then add it again to the branch. The problem with
this is that the adding function (rightfully) increments the changed var
with +1. However we didn't actually add a new process. To fix this
simply decrement the changed variable to reflect that we've also removed
the process from the tree.