Fix race between determining timeout and marking as polling
This problem was highlighted by the example in src/actor_ref/rpc, line
86. However it had nothing to do with RPC or ActorRefs and everything to
do with a race in the use of ThreadWaker.

Before polling we first determine the timeout to use based on local and
shared resources, i.e. the schedulers, timers and wake-up mechanisms.
After we determined the timeout we mark ourselves as polling iff the
timeout is not zero.

However it could be that between those two calls (`determine_timeout`
and `mark_polling`) we received e.g. a wake-up event but didn't consider
it in `determine_timeout`. But because we didn't mark ourselves as
polling yet we also won't be awoken from polling, causing us to poll for
ever, missing the wake-up. That would look something like the following:

Thread 0                    | Thread 1
1. determine_timeout = None |
                            | 2. task::Waker: wake-up pid=1.
                            | 3. ThreadWaker: not polling -> no wakeup.
4. mark_polling(true)       |
5. poll(None)               |
6. Waiting for ever...      |

To fix this we need to check the timeout after marking ourselves as
polling to ensure we didn't miss any wake-ups/timers since we determined
the timeout and marked ourselves as polling.