Improve ThreadWaker::wake performance
By first loading the current status and only after that trying to
atomically change the value. Using Ordering::Relaxed makes the atomic
load a regular load.

It also replaces compare_exchange with fetch_add which I've seen be
faster in the past, but don't have any benchmarks for at the moment.