Optimise Receiver::try_reset
The previous version would attempt to receive the message, moving it
from the heap to the stack and drop it from the stack. However this move
to the stack can be costly for large messages.

This optimised version drops the message in place on heap. The new code
looks similar to that of Receiver::try_recv.