Fix RUSTUP_TARGET var in Makefile
Although it's often referred to as a target-"triple", in some case it's
actually four fields. For example for GNU/Linux the value is
"x86_64-unknown-linux-gnu".

This change include the last part (-gnu) in the RUSTUP_TARGET variable,
but keeps working for targets with only three parts, such as
"x86_64-apple-darwin".