Fix testing of examples with --target specified
The previous version of run_example expected the examples binaries to be
in the target/debug directory, but when --target is specified this is
not the case.

The solution is not pretty, but adds the paths for the supported
targets. It should do this by getting the target triple from
cargo/rustc, but currently this is not possible without a build script.
So the paths are hardcoded based on cfg(target_os).