Add shared Poll instance to shared::RuntimeInternals
This shared Poll instance allows us to register the shared I/O sources
with a single poll, but allows the worker threads to handle the events.
This means that coordinator no longer has to process any events for the
shared processes. This leaves the coordinator to watch over the worker
and sync worker threads.