Don't stop runtime when failing to relay process signal
This removes the coordinator::Error::SignalRelay variant and no longer
stops the runtime (well the coordinator) once a process signal can't be
relayed to a worker thread. Since this can really only happen if the
worker thread is shutdown or in the process of shutting down, it's more
useful to return the reason the worker thread shutdown, i.e. via
rt::Error::worker_panic, rather then a failure to relay the process
signal with an unknown reason.