Remove Worker::handle_messages
Even though it was technically possible to call it served no purpose as
the type of message was ! (the never type), so it couldn't actually
receive any message. It was only there in case the worker thread needed
to communicate with the coordinator, but so far that hasn't been needed,
so we might as well remove the dead code.