Change rt::channel to be a single direction channel
The previous implementation used two pipes and two in-memory channel to
be able to send messages between the coordinator and worker threads.
However so far it's not been needed to send a message from the worker to
the coordinator.

Since it's not needed simple remove on of the pipes and channels,
turning rt::channel into a single direction channel.