Document all usages of atomics
In the past the *correct* use of atomics is trickier than it first
seems. It often requires checking the assembly for correctness and it
requires knowing what other locations access the same memory.

This commit documents all atomic operations that don't use SeqCst why
the ordering doesn't need to be higher and with other point (in the
code) the operations need to synchronise.