Expose lifetime of HeaderName
We'll mostly use a 'static lifetime, e.g. in Headers, but a short
lifetime can be used in various places when we're actually borrowing
data. For example in the Iterator implementation for Headers and
Headers::get, neither now needs to clone the HeaderName and instead lets
the name be borrowed.

This also adds another lifetime to Headers, now 'n and 'v for the name
and value.