Change Connection::{send_response,respond} API
Connection::respond now accepts a reference to the Headers, allowing the
allocating to be reused (see the my_ip example changes).

Connection::send_response now accepts the fields of the Response as
arguments because Headers is owned in it, but accepting a &Response
doesn't work because we do need ownership of the body.