Replace join WakerList with a Vec behind a Mutex
The WakerList was a singly linked list implementation that didn't
properly keep track of the lifetime of the nodes. After various ideas
and designs I came to the conclusion it's not possible to guarantee the
nodes, as stored in the Future struct (i.e. SendValue and Join), always
outlive the linked list. Furthermore slow threads iterating over the
list don't have a way to post progress, thus it's always possible that a
slow thread is accessing a node that is no longer in the list.

The problem(s) described above I've also hit while testing. A slow
thread that iterated over the join wakers and woke them wrote to memory
that was already released (the Join struct was dropped).

In the future we should consider redesigning the wakers lists, for
example by using event-listener (https://docs.rs/event-listener, based
on eventcounts by Dmitry Vyukov), but for now a simple vectored and
mutex will have to do. At least that should work properly.