Add host id to Coordinator metrics
Add a host or machine id to the coordinator metrics.

On Linux this uses the /etc/hostid file, which is a systemd specific
file. More documentation at
https://www.freedesktop.org/software/systemd/man/machine-id.html.

For FreeBSD we use /etc/machine-id, which isn't actually documented
anywhere, but has an open issue to do so:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=255293.

On macOS we can uses the gethostuuid(2) system call.