Implement Clone for EventTiming
This allows the same timing to be used for traces and sub-traces. For
example when you want to tracing the handling of a message you might
want to trace three parts:
* The receiving of a message.
* The handling of a received message.
* The combination of the two about as parent span.

This also adds an example that shows this new usage.

Finally it adds the must_use attribute to Trace::start_trace.