Reduce a10::Ring sizes
The previous sizes we chosen somewhat arbitrarily.

For the coordinator we don't need to many events, so 32 should be
enough, but this might become too small with a large number of worker
threads, so maybe in the future it should be based on the amount of
workers.

For the worker we should poll the ring frequently enough that 64 is
sufficient.

Finally for the shared ring we use 128, which should also be polled
frequently enough.