Update Heph-http
This updates the entire heph-http code base to be up to date with the
v0.5 API and makes it able to compile with the latest nightly.

It updates all dependencies and adds heph-rt as that used be part of the
heph crate, but no longer.

This makes a number of big changes to make everything compile again. But
since they're mostly interconnected it's hard to split them up into
small commits, so I'm just not going to bother (sorry future me).

Changes to the body module
==========================

This changes the design of the Body trait to take ownership of the HTTP
head buffer passed by the caller. This is required by io_uring, but also
makes the lifetime much simpler, which is a nice benefit.

The OneshotBody now accepts a generic type B that must implement the Buf
trait. So we can use owned or (statically) borrowed buffers.

The StreamingBody API is also changed, from Stream to AsyncIterator (to
update rustc version), but the item is changed from io::Result to B:
Buf.

The ChunkedBody API saw the same changes as StreamingBody.

Removed the FileBody type as sendfile isn't supported by io_uring.

Changes to the client module
============================

The Client::connect method now accept a &RT instead of an
actor::Context, matching the changes made to TcpStream::connect and
similar API. Furthermore the method is now an asynchronous function
instead of return the Connect future (which is removed).

The Client::{get, request, read_response} methods now return a single
Result and the I/O errors are moved into the new ResponseError::IO
variant.

Adds Body::chunk_size_hint and is_chunked API.

Removes the Body::read_all method and replaces them with Body::recv and
recv_vectored, which are similar to the TcpStream::recv(_vectored)
methods.

Removes the Copy, Clone and Eq implementations from ResponseError
because I added the IO variant to hold I/O errors. This was required to
allow the Client methods to return a single result.

Changes to the server module
============================

The Server type is now a type alias for heph_rt::nettcp::server::Setup.
The ArgMap struct has been renamed HttpNewActor, which is slighly better
(but still not great).

The actor start by the HTTP server will no longer received the socket
address as that is no longer returned by the underlying TCP server.

The Connection::next_request method now returns a single error, the I/O
error is moved into a RequestError::IO variant, same as we did for
client::ResponseError.

The Body::{recv, recv_vectored} methods are now async functions.