Impl Read trait for various types
This adds implementations to:
 * File
 * Stdin
 * TcpStream
 * UnixStream
 * pipe::Receiver

As well as an implementation for &T for the above types.

This also changes the stdio types, created by the stdio! macro, to use a
`fd: AsyncFd` field rather then a tuple to simplify the impl_read! macro
and be consistent with all the other types.