Rewrite sync worker
Drops the usage of the unix pipe and instead wake the coordinator via
it's a10::Ring, same as what we do for the workers.

I've also rewritten the large main function into a new type SyncWorker.
The main benefit here is to ensure that the waking of the coordinator
always happens (assuming the SyncWorker is dropped properly).

The old SyncWorker type is now called sync_worker::Handle, similar to
the worker::Handle. The only difference between the two is the channel
used by the worker (not sync) threads.