Work around the coordinator wake up data race
When a (sync) worker is stopped it will wake the coordinator in it's
Drop impelementation by a call to shared::RuntimeInternals::wake_coordinator.
This informs the coordinator that the worker is done. However the
coordinator doesn't know which worker is done, so it checks all worker
threads.

A data race exists here between the coordinator waking event (due the
worker's waking) and the coordinator checking the status of that worker
thread. If the coordinator is too fast, or the worker too slow, it cause
the coordinator to check the worker's status before it actually
terminated, making the coordinator think it's still running. If this was
the last worker thread in the coordinator it would mean that the
coordinator will be stuck waiting for ever for event from zero workers
(and process signals).

To work around this, this commits adds a boolean check to make sure that
every wake up we find the reason why and if not found we set a small-ish
timeout for the next poll call to give the worker thread some time to
shutdown before we check again.