Rewrite local waker implementation
This implementation doesn't use static mutable data, but instead
allocates a large mapping of (small) indices to process ids (ProcessId),
which can be squashed into a pointer to this allocation.

This ensures that we don't leak any memory on exit. It also removes the
limit on the number of threads the runtime can use. Furthermore it
removes the use of (mutable) static data.

Future Work

This implement allocates every 2 << 10 (2048) wakers created, but we can
do better if we keep a copy of the waker in the process so that it can
be reused. Of course, this does mean increased memory usage.