Add heph_rt::fs::watch
New module that contains a filesystem change watcher. It's an
implementation based on inotify(7) and expose it's API in straight
forward way. It has some niceties such as recursive directory watching.
However, it being inotify based, means it also has a number of
limitations noted in the documentation of the module.